Manages data loading and pre-processing of the evalPurency package. In the end, also a quality control is performed that throws warnings if there are empty files or fields, including the name and column.

Shared by both main function of the evalPurency package. Loads data from all *.csv files from the provided folter.

Usage

ep.load.helper(
  path,
  particleNumbers,
  sep,
  dec,
  formFillDefault,
  colourFillDefault,
  colL,
  colPol,
  startrow,
  colReqPol,
  colourSep,
  ReqPolKey,
  colShape,
  colCol,
  colLFib,
  colArea,
  colWidth,
  test,
  fibre,
  sphere,
  fragment,
  pixel
)

Arguments

path

The path set by the user in main function.

particleNumbers

set TRUE if you would like to get an extra file with just plastic and non-plastic particle numbers for each file loaded.

sep

Column separator within the *.csv files. Defaults to ';' in main function.

dec

Decimal sign within the *.csv files. Defaults to ',' in main function.

formFillDefault

If desired you can provide a default form that will be filled in, if no form is provided (NA). Should be one of the values of the parameters 'fibre', 'sphere', 'fragment' or 'pixel' (see below).

colourFillDefault

If desired you can provide a default colour that will be filled in, if no colour is provided in the data (NA). However, in this function colour has no further meansing.

colL

Column number for the particle length. In the TOEKI lab this is column 17 (Length 5µ). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d). Set in main function.

colPol

Column number where the polymer type is stated. In the TOEKI lab this is column 6 (Class Name). The polymer names (or abbreviations) can be set manually in 'polymers' in case the default does not apply. Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d). Set in main function.

startrow

Number of rows that can be omitted from the *.csv files. There Purency provides a bunch of meta data, that are not of interest here. Defaults to 40 in main function.

colReqPol

Column number for the particle check, whether the particle is a polymer or not. In the TOEKI lab this is column 24 (Plastik? or Plastik ja/nein). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d). Set in main function.

colourSep

whether colour separation is desired for the analysis. Only usable in evalPurency.particles()!

ReqPolKey

key word or sentence of 'colReqPol' that indicates that it is a plastic particle. Default is 'ja'. Set in main function.

colShape

Column number for the particle shape. In the TOEKI lab this is column 25 (Form). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d). Set in main function.

colCol

Column number for the particle color In the TOEKI lab this is column 26 (Farbe). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d). Set in main function.

colLFib

Column number for the particle length in case of a fibre with corrected length (because of curvy shape) In the TOEKI lab this is column 27 (Länge). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d). Set in main function.

colArea

Column number for the particle length in case of a fibre with corrected length (because of curvy shape) In the TOEKI lab this is column 4 (Area µm²). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d). Set in main function.

colWidth

Column number for the particle length in case of a fibre with corrected length (because of curvy shape) In the TOEKI lab this is column 18 (Width µm). Could also be provided as column name, but only in ASCII encoding (e.g., special character as . and ä = d). Set in main function.

test

Can be set TRUE when the function should be run in testing mode. Set in main function.

fibre

How fibres are called in colShape (Form). In the TOEKI lab it is 'Faser'.

sphere

How spheres are called in colShape (Form). In the TOEKI lab it is 'Kugel'.

fragment

How fragments are called in colShape (Form). In the TOEKI lab it is 'Fragment'.

pixel

How pixels are called in colShape (Form). In the TOEKI lab it is 'Pixel'.

Value

Returns a data frame containing a list of all particles and their properties.